// src/app/core/models/usuario.ts
export interface Usuario {
    id: number;
    nombre: string;
    email: string;
    rol: string;
    fechaRegistro: Date;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  nombre: string;
  email: string;
  password: string;
  rol: string;
}

export interface LoginResponse {
  token: string;
}